Initial voice assistant implementation (#3)
This adds an assistant with the same UI/UX as the swift one:

- a connect button
- a connecting state
- an agent visualizer
- a local audio visualizer and mute button
- a disconnect button

it also includes a TokenService that can fetch tokens from LK sandbox.

Tested on
- [x] iOS
- [x] macOS
- [x] Android

I did test on web but the visualizers don't work so i'm excluding it for
now. I intend to support linux and windows but will address them
separately.

Also still needs

- [x] more testing of mute/unmute
- [x] bootstrap support (luckily flutter already supports `.env.`
files!)
- [x] code cleanup / comments
- [x] Readme / screenshots 
- [x] point to production package versions after updates